Major Changes with an Array
I deleted all of the if statements and replaced with an array in the
Main Window which now holds all of the values in a string for every
parameter. So when we have to go and collect from the data base we can
just check the array instead of checking individual variables.